{"schema":"decodex/commit/1","summary":"Split manual tests","authority":"XY-1115"}
